Safari-609.1.11.1/LayoutTests/storage/websql/open-database-creation-callback-expected.txt

CONSOLE MESSAGE: line 30: Web SQL is deprecated. Please use IndexedDB instead.
CONSOLE MESSAGE: line 56: Web SQL is deprecated. Please use IndexedDB instead.
CONSOLE MESSAGE: line 56: unable to open database, version mismatch, '1.0' does not match the currentVersion of ''
CONSOLE MESSAGE: line 69: Web SQL is deprecated. Please use IndexedDB instead.
CONSOLE MESSAGE: line 70: Web SQL is deprecated. Please use IndexedDB instead.
This test tests openDatabase()'s creation callback.
